auth and security

+++ src/routers/task.js
@@ -1,9 +1,14 @@
 const express = require('express');
 const Task = require('../models/task');
+const auth = require('../middleware/auth')
 const router = new express.Router();
 
-router.post('/tasks', async (request, response)=>{
-  const task = new Task(request.body);
+router.post('/tasks', auth, async (request, response)=>{
+  const task = new Task({
+    ...request.body,
+    owner: request.user._id
+  })
+
   try {
     await task.save();
     response.status(201).send(task);
@@ -13,18 +18,19 @@ router.post('/tasks', async (request, response)=>{
 
 })
 
-router.get('/tasks/', async (request, response)=>{
+router.get('/tasks/', auth, async (request, response)=>{
   try {
-    const tasks = await Task.find({})
-    response.send(tasks)
+    // const tasks = await Task.find({owner: request.user._id});
+    await request.user.populate('tasks').execPopulate()
+    response.send(request.user.tasks)
   } catch(error){
     response.status(500).send(error)
   }
 })
-router.get('/tasks/:id', async (request, response)=>{
+router.get('/tasks/:id', auth, async (request, response)=>{
   const _id = request.params.id;
   try {
-    const task = await Task.findById(_id)
+    const task = await Task.findOne({ _id, owner: request.user._id});
     if(!task) return response.status(404).send();
     response.send(task)
   } catch (error){
@@ -33,9 +39,9 @@ router.get('/tasks/:id', async (request, response)=>{
 })
 
 
-router.delete('/tasks/:id', async (request, response)=>{
+router.delete('/tasks/:id', auth, async (request, response)=>{
   try {
-    const task = await Task.findByIdAndDelete(request.params.id);
+    const task = await Task.findByIdAndDelete({ _id: request.params.id, owner: request.user._id});
     if(!task) return response.status(404).send();
     response.send(task);
   } catch(error){
@@ -45,7 +51,7 @@ router.delete('/tasks/:id', async (request, response)=>{
 })
 
 
-router.patch('/tasks/:id', async (request, response)=>{
+router.patch('/tasks/:id', auth, async (request, response)=>{
   const updates = Object.keys(request.body);
   const allowedUpdates = ['description', 'completed'];
   const isValidOperation = updates.every(
@@ -54,10 +60,13 @@ router.patch('/tasks/:id', async (request, response)=>{
   if(!isValidOperation) return response(400).send({error: 'Invalid updates!'});
 
   try {
-    const task = await Task.findByIdAndUpdate(request.params.id, request.body, {new: true, runValidators: true});
+    const task = await Task.findOne({_id: request.params.id, owner: request.user._id})
     if(!task) return response.status(404).send();
+    updates.forEach((update) => task[update]=request.body[update]);
+    await task.save();
     response.send(task);
   } catch(error){
+    console.log(error);
     response.status(400).send(error)
   }
